+++ apiml-security-common/build.gradle
@@ -1,3 +1,7 @@
+plugins {
+    id "java-test-fixtures"
+}
+
 dependencies {
     api project(':apiml-common')
 
@@ -23,4 +27,13 @@ dependencies {
     testCompileOnly libs.lombok
     testImplementation libs.netty.reactor.http
     testAnnotationProcessor libs.lombok
+
+    testFixturesImplementation libs.nimbus.jose.jwt
+    testFixturesImplementation libs.jose4j.jwt
+    testFixturesImplementation libs.jjwt
+    testFixturesImplementation libs.jjwt.impl
+    testFixturesImplementation libs.jjwt.jackson
+
+    testFixturesCompileOnly libs.lombok
+    testFixturesAnnotationProcessor libs.lombok
 }
